requirements.txt
pandas==2.3.3
numpy==2.3.5
scikit-learn==1.7.2
xgboost
joblib==1.5.2
PyYAML
fastapi==0.136.1
uvicorn
mlflow
prometheus-client
jinja2
pytest
httpx
